FAST -HT-/HT2- refactoring
